requirements.txt
setuptools>=34
wheel
#jcconv
simplejson
Flask
Flask-Babel
Flask-cors
pyusb
pyserial
pif
Pillow
netifaces
pycups
pymtp
git+https://github.com/akretion/py-xml-escpos.git@py3
pypostelium
pyposdisplay
